+++ apps/inspect/src/app/App.tsx
@@ -11,8 +11,8 @@ import "prismjs/components/prism-json";
 import "prismjs/components/prism-python";
 import "prismjs/components/prism-yaml";
 import "prismjs/themes/prism.css";
-import "@tsmono/inspect-common/theme/base";
-import "@tsmono/inspect-common/theme/vscode";
+import "@tsmono/theme/base";
+import "@tsmono/theme/vscode";
 import "./App.css";
 
 import ClipboardJS from "clipboard";
