+++ apps/inspect/src/state/userSettings.ts
@@ -1,7 +1,7 @@
 import { create } from "zustand";
 import { createJSONStorage, persist } from "zustand/middleware";
 
-import type { ThemePreference } from "@tsmono/inspect-common/theme/bootstrap";
+import type { ThemePreference } from "@tsmono/theme/bootstrap";
 
 import { SETTINGS_STORAGE_KEY } from "../theme/constants";
 
@@ -11,7 +11,8 @@ export { SETTINGS_STORAGE_KEY };
 // Mirrors scout's userSettings: a single zustand-persisted store for
 // localStorage-backed UI preferences, ready for more settings to land here
 // rather than each growing its own bespoke record. The persisted key/shape is
-// shared with the inline theme bootstrap (see theme/readPreference.ts).
+// shared with the inline theme bootstrap, which reads it via
+// `readThemePreference` (@tsmono/theme) with this same SETTINGS_STORAGE_KEY.
 interface UserSettingsState {
   themePreference: ThemePreference;
   setThemePreference: (themePreference: ThemePreference) => void;
