+++ apps/scout/package.json
@@ -72,7 +72,6 @@
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^6.0.2",
-    "esbuild": "0.28.0",
     "eslint": "^9.39.4",
     "eventsource": "^4.1.0",
     "jsdom": "^29.1.1",
@@ -100,6 +99,7 @@
     "@tsmono/inspect-components": "workspace:*",
     "@tsmono/react": "workspace:*",
     "@tsmono/scout-components": "workspace:*",
+    "@tsmono/theme": "workspace:*",
     "@tsmono/util": "workspace:*",
     "@uwdata/flechette": "^2.3.0",
     "@vscode-elements/react-elements": "^2.4.0",
