+++ apps/scout/src/app/App.css
@@ -1,4 +1,4 @@
-/* Scout-specific overlay on top of @tsmono/inspect-common/theme/base + @tsmono/inspect-common/theme/vscode.
+/* Scout-specific overlay on top of @tsmono/theme/base + @tsmono/theme/vscode.
    Rules here are either unique to scout or reference assets that live under
    apps/scout/src/img/. Shared rules live in the theme package. */
