+++ packages/inspect-components/src/theme/ThemeToggle.tsx
@@ -1,7 +1,7 @@
 import clsx from "clsx";
 import { useEffect, useId, useState } from "react";
 
-import type { ThemePreference } from "@tsmono/inspect-common/theme/bootstrap";
+import type { ThemePreference } from "@tsmono/theme/bootstrap";
 import { PopOver } from "@tsmono/react/components";
 
 import styles from "./ThemeToggle.module.css";
