+++ packages/theme/eslint.config.js
@@ -1,0 +1,20 @@
+import tseslint from "typescript-eslint";
+
+import baseConfig from "@tsmono/eslint-config/base";
+
+export default tseslint.config(
+  {
+    ignores: ["node_modules/", "dist/", "eslint.config.js"],
+  },
+  ...baseConfig,
+  {
+    languageOptions: {
+      parserOptions: {
+        projectService: {
+          allowDefaultProject: ["eslint.config.js"],
+        },
+        tsconfigRootDir: import.meta.dirname,
+      },
+    },
+  }
+);
