+++ packages/theme/src/vscode.d.ts
