+++ packages/theme/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@tsmono/tsconfig/base.json",
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"]
+}
